requirements.txt
flask==2.3.3
python-dotenv==1.0.0
langchain==0.0.335
langchain-google-genai==0.0.2
google-generativeai==0.3.1
pymupdf==1.23.5
pytesseract==0.3.10
opencv-python-headless==4.8.0.76
numpy==1.24.3
chromadb==0.4.15
sentence-transformers==2.2.2
python-multipart==0.0.6
werkzeug==2.3.7
pillow==10.0.0
